Bugfix: Set invalidation as integer, safe place to come back

+++ include/Entity.hpp
@@ -2,7 +2,7 @@
 
 #include <list>
 #include <string>
-constexpr size_t kInvalidSz = 0xFFFFFFFF;
+constexpr int kInvalidSz = 0xFFFFFFFF;
 
 struct Entity;
 using ListOfEntities = std::list<Entity>;
